test: keep runtime lifecycle gates green
