+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/work_output_materialization.go
@@ -31,7 +31,7 @@ func workstationDispatchResultFromExecute(
 			RetryCount: result.Metrics.RetryCount,
 		},
 		ProviderSession: providerSessionFromContinuation(result.Continuation),
-		Diagnostics:     workers.WorkDiagnosticsFromSafeDiagnostics(result.Diagnostics),
+		Diagnostics:     result.Diagnostics.ToWorkDiagnostics(),
 	}
 	terminal := workers.WorkstationDispatchTerminalOutcomeCompleted
 	switch result.Outcome {
